+++ Eyepatch/memory.h
@@ -8,5 +8,5 @@
 #endif
 
 void* __cdecl operator new(size_t size);
-void* __cdecl operator new(size_t size, POOL_TYPE pool, ULONG tag = EYEPATCH_MEMORY_TAG);
-void __cdecl operator delete(void* p/*, unsigned int dummy*/);
+void* __cdecl operator new(size_t size, POOL_TYPE pool, uint32_t tag = EYEPATCH_MEMORY_TAG);
+void __cdecl operator delete(void* p);
